Revert "feat: Provide support for batch scraping (#2459)"
This reverts commit 5fa547b094949be383e0d5d0c755b173dcc949c0.
