+++ src/Promitor.Core.Scraping/Configuration/Model/Metrics/ScrapeDefinition.cs
@@ -94,16 +94,5 @@ namespace Promitor.Core.Scraping.Configuration.Model.Metrics
             }
             return AzureMetricConfiguration?.Aggregation?.Interval;
         }
-
-        public ScrapeDefinitionBatchProperties BuildScrapingBatchInfo() { 
-            return new ScrapeDefinitionBatchProperties(
-                this.AzureMetricConfiguration,
-                this.LogAnalyticsConfiguration,
-                this.PrometheusMetricDefinition,    
-                this.Resource.ResourceType,
-                this.Scraping,
-                this.SubscriptionId
-            ); 
-         }
     }
 }
