+++ src/Promitor.Core.Scraping/Interfaces/IScraper.cs
@@ -7,6 +7,5 @@ namespace Promitor.Core.Scraping.Interfaces
     public interface IScraper<TResourceDefinition> where TResourceDefinition : class, IAzureResourceDefinition
     {
         Task ScrapeAsync(ScrapeDefinition<TResourceDefinition> scrapeDefinition);
-        Task BatchScrapeAsync(BatchScrapeDefinition<TResourceDefinition> batchScrapeDefinition);
     }
 }
