+++ src/Promitor.Core.Scraping/ResourceTypes/DataShareScraper.cs
@@ -51,7 +51,7 @@ namespace Promitor.Core.Scraping.ResourceTypes
             return new List<string> { dimensionName };
         }
 
-        protected override List<MeasuredMetric> EnrichMeasuredMetrics(DataShareResourceDefinition resourceDefinition, List<string> dimensionNames, IReadOnlyList<MeasuredMetric> metricValues)
+        protected override List<MeasuredMetric> EnrichMeasuredMetrics(DataShareResourceDefinition resourceDefinition, List<string> dimensionNames, List<MeasuredMetric> metricValues)
         {
             // Change Azure Monitor dimension name to more representable value
             foreach (var dimension in metricValues.SelectMany(measuredMetric => measuredMetric.Dimensions.Where(dimension => (dimension.Name == "ShareName" || dimension.Name == "ShareSubscriptionName"))))
@@ -59,7 +59,7 @@ namespace Promitor.Core.Scraping.ResourceTypes
                 dimension.Name = "share_name";
             }
 
-            return metricValues.ToList();
+            return metricValues;
         }
 
         private static string GetMetricFilterFieldName(string metricName)
