+++ src/Promitor.Core.Scraping/Scraper.cs
@@ -98,64 +98,14 @@ namespace Promitor.Core.Scraping
             }
             catch (Exception exception)
             {
-                Logger.LogCritical(exception, "Failed to scrape resource for metric '{MetricName}'. Details: {Details}", scrapeDefinition.PrometheusMetricDefinition.Name, exception.ToString());
+                Logger.LogCritical(exception, "Failed to scrape resource for metric '{MetricName}'", scrapeDefinition.PrometheusMetricDefinition.Name);
 
                 await ReportScrapingOutcomeAsync(scrapeDefinition, isSuccessful: false);
             }
         }
 
-        public async Task BatchScrapeAsync(BatchScrapeDefinition<IAzureResourceDefinition> batchScrapeDefinition)
-        {  
-            // would the large volume of JSON be an issue? Can it be handled by the SDK? 
-            if (batchScrapeDefinition == null)
-            {
-                throw new ArgumentNullException(nameof(batchScrapeDefinition));
-            }
-
-            var aggregationInterval = batchScrapeDefinition.ScrapeDefinitionBatchProperties.GetAggregationInterval();
-            if (aggregationInterval == null)
-            {
-                throw new ArgumentNullException(nameof(batchScrapeDefinition));
-            }
-
-            try
-            {
-                var aggregationType = batchScrapeDefinition.ScrapeDefinitionBatchProperties.AzureMetricConfiguration.Aggregation.Type;
-                var scrapeDefinitions = batchScrapeDefinition.ScrapeDefinitions;
-                var scrapedMetricResults = await BatchScrapeResourceAsync(
-                    batchScrapeDefinition.ScrapeDefinitionBatchProperties.SubscriptionId,
-                    batchScrapeDefinition,
-                    aggregationType,
-                    aggregationInterval.Value);
-                
-                foreach (int i in Enumerable.Range(0, scrapedMetricResults.Count))
-                {
-                    var scrapedMetricResult = scrapedMetricResults[i];
-                    var scrapeDefinition = scrapeDefinitions[i];
-                    LogMeasuredMetrics(scrapeDefinition, scrapedMetricResult, aggregationInterval);
-
-                    await _metricSinkWriter.ReportMetricAsync(scrapeDefinition.PrometheusMetricDefinition.Name, scrapeDefinition.PrometheusMetricDefinition.Description, scrapedMetricResult);
-
-                    await ReportBatchScrapingOutcomeAsync(batchScrapeDefinition, isSuccessful: true, batchScrapeDefinition.ScrapeDefinitions.Count) ;
-                }
-            }
-            catch (ErrorResponseException errorResponseException)
-            {
-                HandleErrorResponseException(errorResponseException, batchScrapeDefinition.ScrapeDefinitionBatchProperties.PrometheusMetricDefinition.Name);
-
-                await ReportBatchScrapingOutcomeAsync(batchScrapeDefinition, isSuccessful: false, batchScrapeDefinition.ScrapeDefinitions.Count);
-            }
-            catch (Exception exception)
-            {
-                Logger.LogCritical(exception, "Failed to scrape resource for metric '{MetricName}'", batchScrapeDefinition.ScrapeDefinitionBatchProperties.AzureMetricConfiguration.MetricName);
-
-                await ReportBatchScrapingOutcomeAsync(batchScrapeDefinition, isSuccessful: false, batchScrapeDefinition.ScrapeDefinitions.Count);
-            }
-        }
-
         private const string ScrapeSuccessfulMetricDescription = "Provides an indication that the scraping of the resource was successful";
         private const string ScrapeErrorMetricDescription = "Provides an indication that the scraping of the resource has failed";
-        private const string BatvhSizeMetricDescription = "Provides an indication that the scraping of the resource has failed";
 
         private async Task ReportScrapingOutcomeAsync(ScrapeDefinition<IAzureResourceDefinition> scrapeDefinition, bool isSuccessful)
         {
@@ -180,46 +130,12 @@ namespace Promitor.Core.Scraping
                 {"resource_group", scrapeDefinition.ResourceGroupName},
                 {"resource_name", scrapeDefinition.Resource.ResourceName},
                 {"resource_type", scrapeDefinition.Resource.ResourceType.ToString()},
-                {"subscription_id", scrapeDefinition.SubscriptionId},
+                {"subscription_id", scrapeDefinition.SubscriptionId}
             };
 
             // Report!
             await AzureScrapingSystemMetricsPublisher.WriteGaugeMeasurementAsync(RuntimeMetricNames.ScrapeSuccessful, ScrapeSuccessfulMetricDescription, successfulMetricValue, labels);
             await AzureScrapingSystemMetricsPublisher.WriteGaugeMeasurementAsync(RuntimeMetricNames.ScrapeError, ScrapeErrorMetricDescription, unsuccessfulMetricValue, labels);
-        }
-
-        private async Task ReportBatchScrapingOutcomeAsync(BatchScrapeDefinition<IAzureResourceDefinition> batchScrapeDefinition, bool isSuccessful, int batchSize)
-        {
-            // We reset all values, by default
-            double successfulMetricValue = 0;
-            double unsuccessfulMetricValue = 0;
-
-            // Based on the result, we reflect that in the metric
-            if (isSuccessful)
-            {
-                successfulMetricValue = 1;
-            }
-            else
-            {
-                unsuccessfulMetricValue = 1;
-            }
-
-            // Enrich with context
-            var labels = new Dictionary<string, string>
-            {
-                {"metric_name", batchScrapeDefinition.ScrapeDefinitionBatchProperties.PrometheusMetricDefinition.Name},
-                {"resource_type", batchScrapeDefinition.ScrapeDefinitionBatchProperties.ResourceType.ToString()},
-                {"subscription_id", batchScrapeDefinition.ScrapeDefinitionBatchProperties.SubscriptionId},
-                {"is_batch", "1"},
-            };
-            // Report!
-            await AzureScrapingSystemMetricsPublisher.WriteGaugeMeasurementAsync(RuntimeMetricNames.ScrapeSuccessful, ScrapeSuccessfulMetricDescription, successfulMetricValue, labels);
-            await AzureScrapingSystemMetricsPublisher.WriteGaugeMeasurementAsync(RuntimeMetricNames.ScrapeError, ScrapeErrorMetricDescription, unsuccessfulMetricValue, labels);
-
-            if (batchSize > 0) 
-            {
-                await AzureScrapingSystemMetricsPublisher.WriteHistogramMeasurementAsync(RuntimeMetricNames.BatchSize, BatvhSizeMetricDescription, batchSize, labels);
-            }
         }
 
         private void LogMeasuredMetrics(ScrapeDefinition<IAzureResourceDefinition> scrapeDefinition, ScrapeResult scrapedMetricResult, TimeSpan? aggregationInterval)
@@ -300,19 +216,6 @@ namespace Promitor.Core.Scraping
             string subscriptionId,
             ScrapeDefinition<IAzureResourceDefinition> scrapeDefinition,
             TResourceDefinition resourceDefinition,
-            PromitorMetricAggregationType aggregationType,
-            TimeSpan aggregationInterval);
-        
-        /// <summary>
-        ///     Scrapes configured resource batch. Should return telemetry for all scrape definitions as a list
-        /// </summary>
-        /// <param name="subscriptionId">Metric subscription Id</param>
-        /// <param name="batchScrapeDefinition">Contains all scrape definitions in the batch and their shared properties(like resource type)</param>
-        /// <param name="aggregationType">Aggregation for the metric to use</param>
-        /// <param name="aggregationInterval">Interval that is used to aggregate metrics</param>
-        protected abstract Task<List<ScrapeResult>> BatchScrapeResourceAsync(
-            string subscriptionId,
-            BatchScrapeDefinition<IAzureResourceDefinition> batchScrapeDefinition,
             PromitorMetricAggregationType aggregationType,
             TimeSpan aggregationInterval);
